rewrite for/doseq to skip chunked path
for/doseq expand inline to code with a chunked and an unchunked path.

Usually only one is exercised by unit tests, so coverage results are
lower than one would expect, given that all the logic *in* the for loop
has been tested.

By removing the chunked path during macro expansion, we can fix this.