Switch over to the indexing reader from tools.reader
Switch over to the indexing push-back reader from tools.reader as it
adds column information to metadata and adds line metadata to lots of
things that the builtin `LineNumberingPushbackReader` misses.

The indexing reader makes reports more accurate and fixes an issue where
if a long form is broken up over multiple lines, some of the lines may
be reported as ignored.