Add hi/lo watermarks & failure thresholds
The high & low watermarks are used to colorize the summary table that is
output to the console. They are configurable by specifying the flags as
follows: --low-watermark=XX and -high-watermark=YY

If the --fail-threshold=ZZ flag is supplied, and the smaller of either
the overall forms percentage or lines percentage, then the build will
be failed with a -3 exit code.