Make form-reader fail with a helpful exception
The cloverage.source/resource-path function fails to find the path for a
namespace if there is a discrepancy between the name of the namespace
and it's filename. Earlier this would result in IllegalArgumentException
thrown by cloverage.source/resource-reader which didn't include any
information about the namespace and hence wasn't much helpful.

This commit modifies the cloverage.source/form-reader function to handle
the above case, and throws a more helpful exception with the namespace
included. Test has been added.