Use Known Data Readers while Reading Forms
While using Cloverage with Clojure CLI, my tagged literals defined in dependencies of my project are not available. However the dynamic var clojure.core/*data-readers* contains the definitions of all tagged literals from all libs because I run Cloverage with the appropriate classpath.

Because Cloverage is using clojure.tools.reader instead of the build-in Clojure reader, the tagged literal definitions have to be bound to clojure.tools.reader/*data-readers*.

This is related to: #255